pyproject.toml
[tool.poetry]
name = "churn-modeling"
version = "0.1.0"
description = "Enterprise-grade customer churn prediction modeling."
authors = ["I. Mert Inan <Alleksenes@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "churn_modeling", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
joblib = "^1.3.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
pydantic = "^2.0.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pyarrow = "^15.0.0"
shap = "^0.45.0"
mlflow = "^2.9.0"
lightgbm = "^4.0.0"
xgboost = "^2.0.0"
optuna = "^3.5.0"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.0"
ipykernel = "^6.25.0"
pytest = "^7.4.0"
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.9.0"
ydata-profiling = "^4.6.0"
pandera = "^0.18.0"
great-expectations = "^0.18.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
